requirements.txt
Django==4.2.25
mysqlclient==2.2.7
Pillow==10.4.0
torch==2.7.1
torchvision==0.22.1
requests==2.32.4
python-dotenv==1.2.1
workos==5.4.0
beautifulsoup4==4.12.3
